# Use memchr accelerated find_byte to scan for matching vowels and consonants