--- a/deny.toml
+++ b/deny.toml
@@ -26,5 +26,7 @@ skip-tree = []
 [sources]
 unknown-registry = "deny"
 unknown-git = "deny"
-allow-registry = [] # This crate has no dependencies
+allow-registry = [
+  "https://github.com/rust-lang/crates.io-index",
+]
 allow-git = []