This PR adds `bstr` as a dependency to get access to high-quality byte manipulation routines.

Depending on `bstr` makes sense given that this crate's intended consumer is `artichoke-backend`  to
implement the `digest` stdlib module. In this context, `bstr` is an acceptable dependency to take since `artichoke-backend` already depends on `bstr`, and likely will continue to depend on `bstr`.

This PR uses `bstr::ByteSlice::find_byte` instead of `iter().position(...)` on the `VOWELS` and `CONSONANTS` arrays. `find_byte` uses a SIMD accelerated `memchr` routine and is much faster.

The API also carries more semantic meaning than what it replaces.

This PR also cleans up some of the decoder logic to use slice patterns where possible.

This PR removes some uses of `u8::try_from` and replaces them with `as` casts since indexes on the `VOWELS` and `CONSONANTS` array are known to be less than `u8::MAX`.