@lopopolo 2020-03-20
# Performance

This commit gives a 6-10% speedup in `decode` performance.

## Before, master @ e25d285

https://github.com/artichoke/bubblebabble/runs/520243709?check_suite_focus=true

```
test decode::benchmark_decode_empty             ... bench:           4 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         241 ns/iter (+/- 71)
test decode::benchmark_decode_vector_pineapple  ... bench:         211 ns/iter (+/- 10)
test encode::benchmark_encode_empty             ... bench:          27 ns/iter (+/- 1)
test encode::benchmark_encode_vector_1234567890 ... bench:         127 ns/iter (+/- 5)
test encode::benchmark_encode_vector_pineapple  ... bench:         109 ns/iter (+/- 2)
```

## After, GH-17 @ c4f3a7b

https://github.com/artichoke/bubblebabble/runs/521027051?check_suite_focus=true

```
est decode::benchmark_decode_empty             ... bench:           4 ns/iter (+/- 0)
test decode::benchmark_decode_vector_1234567890 ... bench:         227 ns/iter (+/- 14)
test decode::benchmark_decode_vector_pineapple  ... bench:         199 ns/iter (+/- 20)
test encode::benchmark_encode_empty             ... bench:          25 ns/iter (+/- 1)
test encode::benchmark_encode_vector_1234567890 ... bench:         139 ns/iter (+/- 7)
test encode::benchmark_encode_vector_pineapple  ... bench:         119 ns/iter (+/- 5)
```